requirements.txt
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-Sock==0.7.0
gunicorn==23.0.0
openai==1.108.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
python-telegram-bot==22.7
websocket-client==1.8.0
-e ../maxbotlib
